requirements.txt
numpy==1.26.4
pandas==1.3.5
Pillow==9.0.1
matplotlib==3.5.1
tensorflow==2.16.1